Refactor : mail,sms 위치 변경
