+++ src/main/java/com/gg/gong9/auth/service/AuthService.java
@@ -12,7 +12,7 @@ import com.gg.gong9.global.exception.exceptions.user.UserExceptionMessage;
 import com.gg.gong9.global.security.cookie.CookieUtil;
 import com.gg.gong9.global.security.jwt.CustomUserDetailsService;
 import com.gg.gong9.global.security.jwt.JwtTokenProvider;
-import com.gg.gong9.mail.service.MailService;
+import com.gg.gong9.notification.mail.service.MailService;
 import com.gg.gong9.user.controller.dto.JoinRequest;
 import com.gg.gong9.user.controller.dto.LoginRequest;
 import com.gg.gong9.user.controller.dto.LoginResponse;
