+++ src/main/java/com/gg/gong9/notification/mail/service/MailService.java
@@ -1,4 +1,4 @@
-package com.gg.gong9.mail.service;
+package com.gg.gong9.notification.mail.service;
 
 public interface MailService {
     // void sendEmail(String to, String subject, String body);
