+++ src/main/java/com/gg/gong9/notification/sms/controller/SmsController.java
@@ -1,6 +1,6 @@
-package com.gg.gong9.notification.controller;
+package com.gg.gong9.notification.sms.controller;
 
-import com.gg.gong9.notification.SmsService.SmsService;
+import com.gg.gong9.notification.sms.service.SmsServiceImpl;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -10,15 +10,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/sms")
 public class SmsController {
 
-    private final SmsService smsService;
+    private final SmsServiceImpl smsService;
 
-    public SmsController(SmsService smsService) {
+    public SmsController(SmsServiceImpl smsService) {
         this.smsService = smsService;
     }
 
     @PostMapping("/send")
     public String sendSms(@RequestParam String to, @RequestParam String message) {
-        smsService.sendSms(to, message);
+        smsService.sendMessage(to, message);
         return "문자 전송 완료!";
     }
 }
