+++ src/main/java/com/gg/gong9/notification/sms/service/SmsService.java
@@ -1,0 +1,5 @@
+package com.gg.gong9.notification.sms.service;
+
+public interface SmsService {
+    void sendMessage(String to, String message);
+}
